fix: REWORK-5 — clippy, fmt, file-size cap, pr-bundle-check
1. clippy map_unwrap_or (rb-rerank/src/llm.rs:72): collapse
   .map(|s| s.clamp(0.0, 1.0)).unwrap_or(0.5) → .map_or(0.5, |s| ...)

2. clippy unnecessary_cast (rb-rerank/src/local.rs:92): remove `as f32`
   (fastembed already returns f32); drop the now-stale #[allow] attribute.

3. fmt alphabetical import order (state.rs:10): move `use rb_rerank::Reranker`
   after `rb_query::ModuleTreeCache` per rustfmt ordering.

4. fmt line wrapping: cargo fmt fixed dispatch.rs and llm.rs chain formatting.

5. file-size cap (600-line limit):
   - search.rs (612→435 lines): extract #[cfg(test)] mod tests to
     routes/query/search_tests.rs via #[path].
   - integration_ingest_tests.rs (603→382 lines): move ingest regression test
     to tests/integration_rusaa_ingest_tests.rs (self-contained with helper).

6. pr-bundle-check: amended preceding commit to remove stale tracker
   reference from body (that issue belongs to a different PR, not S7).

Production code and Gates 1-2 remain approved from prior rounds.
