+++ services/control-api/src/state.rs
@@ -7,11 +7,11 @@ use std::time::{Duration, Instant};
 use chrono::{DateTime, Utc};
 use dashmap::DashMap;
 use rb_auth::{LoginRateLimiter, PasswordHasher};
-use rb_rerank::Reranker;
 use rb_email::EmailSender;
 use rb_github::GhAppLoader;
 use rb_kafka::Producer;
 use rb_query::ModuleTreeCache;
+use rb_rerank::Reranker;
 use rb_schemas::{AgentSessionCommand, IngestRequest, Tombstone};
 use rb_sse::EventBus;
 use rb_storage_neo4j::TenantGraph;
