+++ services/control-api/tests/integration_ingest_tests.rs
@@ -379,225 +379,3 @@ async fn ac6_trigger_rolls_back_db_on_kafka_failure() {
         "AC6: pipeline_stage_runs must be absent after Kafka publish failure"
     );
 }
-
-// ---------------------------------------------------------------------------
-// RUSAA-1560 — finished_at must advance to MAX(stage.finished_at)
-// ---------------------------------------------------------------------------
-
-/// AC: `ingestion_runs.finished_at ≥ MAX(pipeline_stage_runs.finished_at)` across
-/// all 9 stages.
-///
-/// Verifies the two-UPDATE logic in `maybe_complete_run`:
-/// 1. Initial completion: `finished_at` is set to MAX of stage timestamps, not `now()`.
-/// 2. Subsequent fan-out Done events: `finished_at` advances as later stages complete.
-///
-/// This test does NOT call Kafka; it exercises the DB queries directly to confirm
-/// the SQL semantics of the fix.
-#[tokio::test]
-#[allow(clippy::too_many_lines)]
-async fn rusaa_1560_finished_at_equals_max_stage_finished_at() {
-    let Some((_state, pool)) = real_db_state().await else {
-        return; // skip: no DB
-    };
-
-    // --- Setup minimal rows ---
-    let tenant_id = Uuid::new_v4();
-    let user_id = Uuid::new_v4();
-    let repo_id = Uuid::new_v4();
-    let run_id = Uuid::new_v4();
-
-    let slug = format!("rusaa1560-{}", tenant_id.simple());
-    let schema_name = format!("rusaa1560_{}", tenant_id.simple());
-
-    sqlx::query(
-        "INSERT INTO control.tenants (id, slug, name, schema_name) VALUES ($1, $2, $3, $4)",
-    )
-    .bind(tenant_id)
-    .bind(&slug)
-    .bind("RUSAA-1560 Test Tenant")
-    .bind(&schema_name)
-    .execute(&pool)
-    .await
-    .expect("insert tenant");
-
-    sqlx::query(
-        "INSERT INTO control.users (id, email, password_hash, email_verified_at) \
-         VALUES ($1, $2, $3, now())",
-    )
-    .bind(user_id)
-    .bind(format!("rusaa1560-{}@test.example", user_id.simple()))
-    .bind("$argon2id$v=19$m=65536,t=1,p=1$placeholder_hash")
-    .execute(&pool)
-    .await
-    .expect("insert user");
-
-    sqlx::query(
-        "INSERT INTO control.tenant_members (tenant_id, user_id, role) VALUES ($1, $2, 'owner')",
-    )
-    .bind(tenant_id)
-    .bind(user_id)
-    .execute(&pool)
-    .await
-    .expect("insert tenant_member");
-
-    sqlx::query(
-        "INSERT INTO control.repos \
-         (id, tenant_id, github_repo_id, full_name, default_branch, connected_by) \
-         VALUES ($1, $2, $3, $4, $5, $6)",
-    )
-    .bind(repo_id)
-    .bind(tenant_id)
-    .bind(42_i64)
-    .bind("test-org/rusaa1560-repo")
-    .bind("main")
-    .bind(user_id)
-    .execute(&pool)
-    .await
-    .expect("insert repo");
-
-    sqlx::query(
-        "INSERT INTO control.ingestion_runs (id, tenant_id, repo_id, status, requested_by) \
-         VALUES ($1, $2, $3, 'running', $4)",
-    )
-    .bind(run_id)
-    .bind(tenant_id)
-    .bind(repo_id)
-    .bind(user_id)
-    .execute(&pool)
-    .await
-    .expect("insert ingestion_run");
-
-    // Insert all 9 stages with succeeded status. Serial stages get early timestamps;
-    // fan-out stages (extract onwards) get progressively later ones — simulating that
-    // they finish long after the serial chain completes.
-    let stages: &[(&str, &str)] = &[
-        ("clone", "now() - interval '300 seconds'"),
-        ("expand", "now() - interval '299 seconds'"),
-        ("parse", "now() - interval '298 seconds'"),
-        ("typecheck", "now() - interval '291 seconds'"), // serial chain done ~9s in
-        ("extract", "now() - interval '33 seconds'"),    // fan-out: first item early
-        ("embed", "now() - interval '27 seconds'"),
-        ("project_pg", "now() - interval '17 seconds'"),
-        ("project_neo4j", "now() - interval '4 seconds'"),
-        ("project_qdrant", "now() - interval '2 seconds'"), // last to finish
-    ];
-    for (stage, ts_expr) in stages {
-        let sql = format!(
-            "INSERT INTO control.pipeline_stage_runs \
-             (id, ingestion_run_id, stage, status, finished_at) \
-             VALUES (gen_random_uuid(), $1, $2, 'succeeded', {ts_expr})"
-        );
-        sqlx::query(&sql)
-            .bind(run_id)
-            .bind(*stage)
-            .execute(&pool)
-            .await
-            .unwrap_or_else(|e| panic!("insert stage {stage}: {e}"));
-    }
-
-    // --- Execute the two UPDATEs from maybe_complete_run ---
-    sqlx::query(
-        "UPDATE control.ingestion_runs \
-         SET status = 'succeeded' \
-         WHERE id = $1 AND status IN ('queued', 'running')",
-    )
-    .bind(run_id)
-    .execute(&pool)
-    .await
-    .expect("status transition");
-
-    sqlx::query(
-        "UPDATE control.ingestion_runs \
-         SET finished_at = (\
-           SELECT MAX(psr.finished_at) \
-           FROM control.pipeline_stage_runs psr \
-           WHERE psr.ingestion_run_id = $1\
-         ) \
-         WHERE id = $1 AND status = 'succeeded'",
-    )
-    .bind(run_id)
-    .execute(&pool)
-    .await
-    .expect("advance finished_at");
-
-    // --- Verify AC: finished_at == MAX(stage.finished_at) ---
-    let (run_finished_at, max_stage_finished_at): (
-        Option<chrono::DateTime<chrono::Utc>>,
-        Option<chrono::DateTime<chrono::Utc>>,
-    ) = sqlx::query_as(
-        "SELECT ir.finished_at, \
-                (SELECT MAX(psr.finished_at) FROM control.pipeline_stage_runs psr \
-                 WHERE psr.ingestion_run_id = ir.id) \
-         FROM control.ingestion_runs ir \
-         WHERE ir.id = $1",
-    )
-    .bind(run_id)
-    .fetch_one(&pool)
-    .await
-    .expect("fetch run + max stage time");
-
-    let run_ts = run_finished_at.expect("finished_at must not be NULL after completion");
-    let max_ts = max_stage_finished_at.expect("at least one stage must have a finished_at");
-
-    assert_eq!(
-        run_ts, max_ts,
-        "RUSAA-1560: ingestion_runs.finished_at must equal MAX(pipeline_stage_runs.finished_at); \
-         got run={run_ts}, max_stage={max_ts}"
-    );
-
-    // --- Simulate a second fan-out Done event (later item) ---
-    // Advance project_qdrant to the LATEST timestamp, representing the last item processed.
-    sqlx::query(
-        "UPDATE control.pipeline_stage_runs \
-         SET finished_at = now() \
-         WHERE ingestion_run_id = $1 AND stage = 'project_qdrant'",
-    )
-    .bind(run_id)
-    .execute(&pool)
-    .await
-    .expect("advance project_qdrant finished_at");
-
-    // Re-run the finished_at advance (as maybe_complete_run would after the next Done event).
-    sqlx::query(
-        "UPDATE control.ingestion_runs \
-         SET finished_at = (\
-           SELECT MAX(psr.finished_at) \
-           FROM control.pipeline_stage_runs psr \
-           WHERE psr.ingestion_run_id = $1\
-         ) \
-         WHERE id = $1 AND status = 'succeeded'",
-    )
-    .bind(run_id)
-    .execute(&pool)
-    .await
-    .expect("re-advance finished_at after later item");
-
-    let (new_run_ts, new_max_ts): (
-        Option<chrono::DateTime<chrono::Utc>>,
-        Option<chrono::DateTime<chrono::Utc>>,
-    ) = sqlx::query_as(
-        "SELECT ir.finished_at, \
-                (SELECT MAX(psr.finished_at) FROM control.pipeline_stage_runs psr \
-                 WHERE psr.ingestion_run_id = ir.id) \
-         FROM control.ingestion_runs ir \
-         WHERE ir.id = $1",
-    )
-    .bind(run_id)
-    .fetch_one(&pool)
-    .await
-    .expect("re-fetch run + max stage time");
-
-    let new_run_ts = new_run_ts.expect("finished_at still non-null after second event");
-    let new_max_ts = new_max_ts.expect("max stage still non-null");
-
-    assert!(
-        new_run_ts > run_ts,
-        "RUSAA-1560: finished_at must advance when a later fan-out item arrives; \
-         before={run_ts}, after={new_run_ts}"
-    );
-    assert_eq!(
-        new_run_ts, new_max_ts,
-        "RUSAA-1560: finished_at must equal MAX(stage.finished_at) after second advance; \
-         got run={new_run_ts}, max_stage={new_max_ts}"
-    );
-}
